requirements.txt
annotated-types==0.7.0
anthropic==0.45.0
anyio==4.8.0
black==24.10.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cohere==5.13.11
colorama==0.4.6
distro==1.9.0
eval_type_backport==0.2.2
fastavro==1.10.0
filelock==3.17.0
fsspec==2024.12.0
google-auth==2.38.0
griffe==1.5.5
groq==0.15.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
jiter==0.8.2
jsonpath-python==1.0.6
Kurigram @ https://github.com/KurimuzonAkuma/pyrogram/archive/dev.zip
logfire-api==3.3.0
mistralai==1.4.0
mypy-extensions==1.0.0
openai==1.60.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclean==3.1.0
pydantic==2.10.6
pydantic-ai==0.0.20
pydantic-ai-slim==0.0.20
pydantic-graph==0.0.20
pydantic-settings==2.7.1
pydantic_core==2.27.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
TgCrypto==1.2.5
tokenizers==0.21.0
tqdm==4.67.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvloop==0.21.0
